Added this test for other databases
Refactor inferNullability test logic

Extracted common inferNullability test logic to `commonTestScenarios.kt` for reusability. Removed redundant code from individual test files and added necessary imports to support the new structure.
